:art: Move doc and settings files in jpyltime packag
